+++ src/vs/workbench/contrib/tasks/browser/taskQuickPick.ts
@@ -7,7 +7,6 @@ import * as nls from 'vs/nls';
 import { Task, ContributedTask, CustomTask, ConfiguringTask, TaskSorter } from 'vs/workbench/contrib/tasks/common/tasks';
 import { IWorkspace, IWorkspaceFolder } from 'vs/platform/workspace/common/workspace';
 import * as Types from 'vs/base/common/types';
-import { URI } from 'vs/base/common/uri';
 import { ITaskService, WorkspaceFolderTaskResult } from 'vs/workbench/contrib/tasks/common/taskService';
 import { IQuickPickItem, QuickPickInput, IQuickPick } from 'vs/base/parts/quickinput/common/quickInput';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
@@ -22,55 +21,10 @@ export function isWorkspaceFolder(folder: IWorkspace | IWorkspaceFolder): folder
 	return 'uri' in folder;
 }
 
-export class TaskMap {
-	private _store: Map<string, Task[]> = new Map();
-
-	public forEach(callback: (value: Task[], folder: string) => void): void {
-		this._store.forEach(callback);
-	}
-
-	private getKey(workspaceFolder: IWorkspace | IWorkspaceFolder | string): string {
-		let key: string | undefined;
-		if (Types.isString(workspaceFolder)) {
-			key = workspaceFolder;
-		} else {
-			const uri: URI | null | undefined = isWorkspaceFolder(workspaceFolder) ? workspaceFolder.uri : workspaceFolder.configuration;
-			key = uri ? uri.toString() : '';
-		}
-		return key;
-	}
-
-	public get(workspaceFolder: IWorkspace | IWorkspaceFolder | string): Task[] {
-		const key = this.getKey(workspaceFolder);
-		let result: Task[] | undefined = this._store.get(key);
-		if (!result) {
-			result = [];
-			this._store.set(key, result);
-		}
-		return result;
-	}
-
-	public add(workspaceFolder: IWorkspace | IWorkspaceFolder | string, ...task: Task[]): void {
-		const key = this.getKey(workspaceFolder);
-		let values = this._store.get(key);
-		if (!values) {
-			values = [];
-			this._store.set(key, values);
-		}
-		values.push(...task);
-	}
-
-	public all(): Task[] {
-		let result: Task[] = [];
-		this._store.forEach((values) => result.push(...values));
-		return result;
-	}
-}
-
 export interface TaskQuickPickEntry extends IQuickPickItem {
 	task: Task | undefined | null;
 }
-export interface TaskQuickPickEntry2 extends IQuickPickItem {
+export interface TaskTwoLevelQuickPickEntry extends IQuickPickItem {
 	task: Task | ConfiguringTask | string | undefined | null;
 }
 
@@ -90,25 +44,25 @@ export class TaskQuickPick extends Disposable {
 		return this.configurationService.getValue<boolean>(QUICKOPEN_DETAIL_CONFIG);
 	}
 
-	private createTaskEntry(task: Task | ConfiguringTask): TaskQuickPickEntry2 {
+	private createTaskEntry(task: Task | ConfiguringTask): TaskTwoLevelQuickPickEntry {
 		let entryLabel = task._label;
 		let commonKey = task._id.split('|');
 		if (commonKey.length > 1) {
 			entryLabel = entryLabel + ' (' + commonKey[1] + ')';
 		}
-		const entry: TaskQuickPickEntry2 = { label: entryLabel, description: this.taskService.getTaskDescription(task), task, detail: this.showDetail() ? task.configurationProperties.detail : undefined };
+		const entry: TaskTwoLevelQuickPickEntry = { label: entryLabel, description: this.taskService.getTaskDescription(task), task, detail: this.showDetail() ? task.configurationProperties.detail : undefined };
 		entry.buttons = [{ iconClass: 'codicon-gear', tooltip: nls.localize('configureTask', "Configure Task") }];
 		return entry;
 	}
 
-	private createEntriesForGroup(entries: QuickPickInput<TaskQuickPickEntry2>[], tasks: (Task | ConfiguringTask)[], groupLabel: string) {
+	private createEntriesForGroup(entries: QuickPickInput<TaskTwoLevelQuickPickEntry>[], tasks: (Task | ConfiguringTask)[], groupLabel: string) {
 		entries.push({ type: 'separator', label: groupLabel });
 		tasks.forEach(task => {
 			entries.push(this.createTaskEntry(task));
 		});
 	}
 
-	private createTypeEntries(entries: QuickPickInput<TaskQuickPickEntry2>[], types: string[]) {
+	private createTypeEntries(entries: QuickPickInput<TaskTwoLevelQuickPickEntry>[], types: string[]) {
 		entries.push({ type: 'separator', label: nls.localize('contributedTasks', "contributed") });
 		types.forEach(type => {
 			entries.push({ label: type, task: type, description: SHOW_ALL });
@@ -148,11 +102,11 @@ export class TaskQuickPick extends Disposable {
 		return dedupedConfiguredTasks;
 	}
 
-	private async createTopLevelEntries(defaultEntry?: TaskQuickPickEntry): Promise<{ entries: QuickPickInput<TaskQuickPickEntry2>[], isSingleConfigured?: Task | ConfiguringTask }> {
+	private async createTopLevelEntries(defaultEntry?: TaskQuickPickEntry): Promise<{ entries: QuickPickInput<TaskTwoLevelQuickPickEntry>[], isSingleConfigured?: Task | ConfiguringTask }> {
 		const recentTasks: (Task | ConfiguringTask)[] = (await this.taskService.readRecentTasks()).reverse();
 		const configuredTasks: (Task | ConfiguringTask)[] = this.handleFolderTaskResult(await this.taskService.getWorkspaceTasks());
 		const extensionTaskTypes = this.taskService.taskTypes();
-		const taskQuickPickEntries: QuickPickInput<TaskQuickPickEntry2>[] = [];
+		const taskQuickPickEntries: QuickPickInput<TaskTwoLevelQuickPickEntry>[] = [];
 		if (recentTasks.length > 0) {
 			this.createEntriesForGroup(taskQuickPickEntries, recentTasks, nls.localize('recentlyUsed', 'recently used'));
 		}
@@ -175,7 +129,7 @@ export class TaskQuickPick extends Disposable {
 	}
 
 	private async show(placeHolder: string, defaultEntry?: TaskQuickPickEntry): Promise<Task | undefined | null> {
-		const picker: IQuickPick<TaskQuickPickEntry2> = this.quickInputService.createQuickPick();
+		const picker: IQuickPick<TaskTwoLevelQuickPickEntry> = this.quickInputService.createQuickPick();
 		picker.placeholder = placeHolder;
 		picker.matchOnDescription = true;
 		picker.ignoreFocusOut = false;
@@ -197,7 +151,7 @@ export class TaskQuickPick extends Disposable {
 			picker.dispose();
 			return this.toTask(topLevelEntriesResult.isSingleConfigured);
 		}
-		const taskQuickPickEntries: QuickPickInput<TaskQuickPickEntry2>[] = topLevelEntriesResult.entries;
+		const taskQuickPickEntries: QuickPickInput<TaskTwoLevelQuickPickEntry>[] = topLevelEntriesResult.entries;
 
 		do {
 			const firstLevelTask = await this.doPickerFirstLevel(picker, taskQuickPickEntries);
@@ -221,9 +175,9 @@ export class TaskQuickPick extends Disposable {
 		return;
 	}
 
-	private async doPickerFirstLevel(picker: IQuickPick<TaskQuickPickEntry2>, taskQuickPickEntries: QuickPickInput<TaskQuickPickEntry2>[]): Promise<Task | ConfiguringTask | string | null | undefined> {
+	private async doPickerFirstLevel(picker: IQuickPick<TaskTwoLevelQuickPickEntry>, taskQuickPickEntries: QuickPickInput<TaskTwoLevelQuickPickEntry>[]): Promise<Task | ConfiguringTask | string | null | undefined> {
 		picker.items = taskQuickPickEntries;
-		const firstLevelPickerResult = await new Promise<TaskQuickPickEntry2 | undefined | null>(resolve => {
+		const firstLevelPickerResult = await new Promise<TaskTwoLevelQuickPickEntry | undefined | null>(resolve => {
 			Event.once(picker.onDidAccept)(async () => {
 				resolve(picker.selectedItems ? picker.selectedItems[0] : undefined);
 			});
@@ -231,7 +185,7 @@ export class TaskQuickPick extends Disposable {
 		return firstLevelPickerResult?.task;
 	}
 
-	private async doPickerSecondLevel(picker: IQuickPick<TaskQuickPickEntry2>, type: string) {
+	private async doPickerSecondLevel(picker: IQuickPick<TaskTwoLevelQuickPickEntry>, type: string) {
 		picker.busy = true;
 		picker.value = '';
 		if (type === SHOW_ALL) {
@@ -240,7 +194,7 @@ export class TaskQuickPick extends Disposable {
 			picker.items = await this.getEntriesForProvider(type);
 		}
 		picker.busy = false;
-		const secondLevelPickerResult = await new Promise<TaskQuickPickEntry2 | undefined | null>(resolve => {
+		const secondLevelPickerResult = await new Promise<TaskTwoLevelQuickPickEntry | undefined | null>(resolve => {
 			Event.once(picker.onDidAccept)(async () => {
 				resolve(picker.selectedItems ? picker.selectedItems[0] : undefined);
 			});
@@ -249,9 +203,9 @@ export class TaskQuickPick extends Disposable {
 		return secondLevelPickerResult;
 	}
 
-	private async getEntriesForProvider(type: string): Promise<QuickPickInput<TaskQuickPickEntry2>[]> {
+	private async getEntriesForProvider(type: string): Promise<QuickPickInput<TaskTwoLevelQuickPickEntry>[]> {
 		const tasks = (await this.taskService.tasks({ type })).sort((a, b) => this.sorter.compare(a, b));
-		let taskQuickPickEntries: QuickPickInput<TaskQuickPickEntry2>[];
+		let taskQuickPickEntries: QuickPickInput<TaskTwoLevelQuickPickEntry>[];
 		if (tasks.length > 0) {
 			taskQuickPickEntries = tasks.map(task => this.createTaskEntry(task));
 			taskQuickPickEntries.unshift({
@@ -275,7 +229,7 @@ export class TaskQuickPick extends Disposable {
 		return this.taskService.tryResolveTask(task);
 	}
 
-	static async show(taskService: ITaskService, configurationService: IConfigurationService, quickInputService: IQuickInputService, placeHolder: string, defaultEntry?: TaskQuickPickEntry, selectedEntry?: TaskQuickPickEntry, additionalEntries?: TaskQuickPickEntry[], tasks?: Promise<Task[]> | Task[], taskSet?: TaskMap) {
+	static async show(taskService: ITaskService, configurationService: IConfigurationService, quickInputService: IQuickInputService, placeHolder: string, defaultEntry?: TaskQuickPickEntry) {
 		const taskQuickPick = new TaskQuickPick(taskService, configurationService, quickInputService);
 		return taskQuickPick.show(placeHolder, defaultEntry);
 	}
